Move Temperature Menu to its own file
